+++ src/lib.rs
@@ -1,9 +1,9 @@
 #[macro_use]
 extern crate nom;
 
+pub mod color_parser;
 pub mod constant;
 pub mod die;
 pub mod operation;
 pub mod operators;
-pub mod parser;
 pub mod traits;
